+++ atous_sec_network/security/security_presets.py
@@ -6,43 +6,9 @@ allowing easy switching between development, staging, and production security le
 
 from dataclasses import dataclass
 from typing import List, Optional
-from .security_middleware import RateLimitConfig
+from .config import RateLimitConfig, SecurityPreset
 
-@dataclass
-class SecurityPreset:
-    """Security configuration preset for a specific environment"""
-    name: str
-    description: str
-    
-    # Rate limiting
-    rate_limit_config: RateLimitConfig
-    
-    # DDoS protection
-    max_connections_per_ip: int
-    connection_timeout_seconds: int
-    
-    # Input validation
-    enable_input_validation: bool
-    strict_validation: bool
-    block_suspicious_patterns: bool
-    
-    # Request limits
-    max_request_size_mb: int
-    max_query_params: int
-    max_header_size: int
-    
-    # Auto-blocking
-    auto_block_malicious_threshold: int
-    auto_block_suspicious_threshold: int
-    block_duration_hours: int
-    
-    # Excluded paths (paths that bypass strict validation)
-    excluded_paths: List[str]
-    
-    # Security headers
-    strict_security_headers: bool
-    enable_csp: bool
-    enable_hsts: bool
+
 
 # Development preset - More permissive for testing
 DEV_PRESET = SecurityPreset(
